refactor: 좋아요 버튼 커스텀 훅으로 관리
